카페 상세 정보 가져오기

+++ src/main/java/com/finalProject/Back/repository/ReviewMapper.java
@@ -2,14 +2,14 @@ package com.finalProject.Back.repository;
 
 import com.finalProject.Back.dto.response.RespReviewInfoDto;
 import com.finalProject.Back.entity.Cafe.CafeDetail;
-import com.finalProject.Back.entity.Review;
+import com.finalProject.Back.entity.Review.Review;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
 
 @Mapper
 public interface ReviewMapper {
-    int save(Review review);
+    Long save(Review review);
     CafeDetail findByCafeId(Long cafeId);
     int getCountByCafeId(Long cafeId);
     int modify(Review review);
